fix: 🐛 update last_login when auth finish
