+++ wangwang/account/signals.py
@@ -1,0 +1,3 @@
+from django.dispatch import Signal
+
+user_logged_in = Signal(providing_args=['user'])
